+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+### Changed
+
+- BREAKING: The error interface no longer extends `Stringable`. Use `$error->message()` instead, or compose a string
+  from multiple properties of the error object.
+- BREAKING: The code on an error object is now type-hinted as a `BackedEnum` or `null` - previously it was `mixed`.
+  Error codes should be from a defined list, therefore an enum is the correctly defined type.
+
 ## [0.1.0] - 2023-11-18
 
 Initial release.